Moved store creation logic into "getStore" function
